move device registry and qserver utils up and merge into single file
