--- server/utils/__init__.py
@@ -1,1 +1,0 @@
-# Utils package for Ophyd as a Service
